fix: repository function decoupled with business action demands

+++ src/core/use-cases/board/delete-board.usecase.ts
@@ -1,3 +1,4 @@
+import { NotBoardOwnerError } from "../../errors/board.error";
 import type { BoardActionEmitter } from "../../interfaces/emitter/board-action-emitter.interface";
 import type { BoardRepository } from "../../interfaces/repo/board-repository.interface";
 import type { MemberRepository } from "../../interfaces/repo/member-repository.interface";
@@ -12,15 +13,14 @@ export class DeleteBoard {
   ) {}
 
   execute = async (boardId: string, userId: string) => {
-    const board = await this.boardRepo.getByIdAfterEnsuringOwner(
-      boardId,
-      userId,
-    );
+    const board = await this.boardRepo.getById(boardId);
+    if (board.attrbs.ownerId !== userId) throw new NotBoardOwnerError();
+
     const members = await this.memberRepo.getAllMembersOfBoardById(boardId);
 
     await this.uow.withTransaction(async () => {
       await this.boardRepo.remove(board);
-      await this.memberRepo.removeAllMembersOfBoard(board);
+      await this.memberRepo.removeAllBoardMembers(board);
     });
 
     this.boardActionEmit.emitBoardDeleted(members, board);
