+++ src/core/use-cases/column/remove-column.usecase.ts
@@ -1,3 +1,4 @@
+import { ColumnNotInBoardError } from "../../errors/column.error";
 import type { ColumnActionEmitter } from "../../interfaces/emitter/column-action-emitter.interface";
 import type { MemberPolicy } from "../../interfaces/policy/member-policy.interface";
 import type { ColumnRepository } from "../../interfaces/repo/column-repository.interface";
@@ -14,10 +15,9 @@ export class RemoveColumn {
   execute = async (columnId: string, boardId: string, userId: string) => {
     await this.memberPolicy.ensureMember(userId, boardId);
 
-    const column = await this.columnRepo.getByIdAfterEnsuringInBoard(
-      columnId,
-      boardId,
-    );
+    const column = await this.columnRepo.getById(columnId);
+    if (column.attrbs.boardId !== boardId) throw new ColumnNotInBoardError();
+
     await this.columnRepo.remove(column);
 
     const members = await this.memberRepo.getAllMembersOfBoardById(boardId);
